chore: move root reducer to src root
